began changing structure of spotify objects
